requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
PyPDF2==3.0.1
python-docx==1.1.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
openai==1.52.2